Removing CommonUtils again
